move VersionedName to util package
