+++ src/main/java/net/fabricmc/tinyremapper/util/VersionedName.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package net.fabricmc.tinyremapper;
+package net.fabricmc.tinyremapper.util;
 
 import java.io.File;
 import java.util.Objects;
